refactor(ring_transport): split transporter.rs by extracting manager + tests (#244)
transporter.rs was 769 lines (over the 700-line cap, including tests).
Split into a directory:

  transporter/mod.rs    — RingTransporter FSM + Effect/State/Timers/
                          RegionEvent/DeadlineKind enums + timing constants
  transporter/manager.rs — RingTransporterManager + RawDeadline (the
                           process-wide registry shim)
  transporter/tests.rs  — existing FSM unit tests, unchanged behavior

The author's prior note about not splitting (the "manager out + Effect
out produces uneven files" case) was conditioned on tests staying inline.
With tests extracted, manager + tests together pull ~260 lines out of
the FSM file, dropping the production-code module to ~445 lines (under
the 500 soft cap).

Public surface preserved: parent declares `pub use transporter::{Effect,
RegionEvent, RingTransporter, RingTransporterManager, State}` — `mod.rs`
re-exports `RawDeadline` and `RingTransporterManager` from the manager
sub-module so the parent line still resolves.

All 8 transporter tests pass; clippy clean; cargo fmt applied.
